put templates and app back in root folder
