+++ db_create_users.py
@@ -1,5 +1,5 @@
-from app import db
-from models import User
+from project import db
+from project.models import User
 
 # insert data
 db.session.add(User("admin", "ad@min.com", "password"))
